+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 All notable changes to `pdf-to-image` will be documented in this file
 
+## Unreleased
+
+- Silently ignore failure of Imagick `pingImage` method
+- Lazy load number of pages to speed up usage when not needed
 
 ## 2.2.0 - 2022-03-08